# services/api/internal/apidocs/openapi.yaml
openapi: 3.0.3

info:
  title: Безопасная сделка — API антискам-тренажёра
  version: "1.0.0"
  description: |
    Бэкенд обучающего тренажёра: профиль пользователя, пошаговое обучение и экзамен —
    непрерывный диалог с языковой моделью, играющей мошенника.

    **Идентификация.** Пользователь анонимный, опознаётся по httpOnly-куке `antiscam_session`.
    Кука ставится ответом `POST /users`. Клиент её не читает и не передаёт в теле.

    **Базовый путь.** Все ручки доступны под `/api/v1` и под алиасом `/api` без версии.
    Голые пути вида `/me` через nginx уходят в SPA и возвращают HTML со статусом 200,
    а не 404 — на бэкенд попадает только то, что начинается с `/api/`.

servers:
  - url: /api/v1
    description: Основной путь
  - url: /api
    description: Алиас без версии

tags:
  - name: Пользователь
  - name: Обучение
  - name: Экзамен
  - name: Результаты
  - name: Служебные

paths:
  /me:
    get:
      tags: [Пользователь]
      summary: Профиль и прогресс по обеим ролям
      description: |
        Если куки нет или пользователь не найден — `exists: false`, `user: null` и статус 200.
        Отсутствие профиля не считается ошибкой.
      responses:
        "200":
          description: Профиль или его отсутствие
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
              examples:
                новый:
                  summary: Аноним без профиля
                  value: { exists: false, user: null }
                существующий:
                  summary: Профиль заведён
                  value:
                    exists: true
                    user:
                      name: Дмитрий
                      age: "25-34"
                      gender: male
                      buyer:
                        training: { currentStep: 6, totalSteps: 6 }
                        isExamPassed: true
                        status: exam_passed
                        isTrainingPassed: true
                      seller:
                        training: { currentStep: 0, totalSteps: 6 }
                        isExamPassed: false
                        status: not_started
                        isTrainingPassed: false

  /users:
    post:
      tags: [Пользователь]
      summary: Создать или обновить профиль
      description: |
        Ставит httpOnly-куку сессии и заводит записи прогресса по обеим ролям.
        Если кука уже есть — профиль обновляется, прогресс не сбрасывается.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, age, gender]
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: Дмитрий
                age:
                  type: string
                  description: Свободная строка, закрытого списка нет
                  example: "25-34"
                gender:
                  $ref: "#/components/schemas/Gender"
      responses:
        "200":
          description: Профиль создан или обновлён, кука поставлена
          headers:
            Set-Cookie:
              schema:
                type: string
                example: antiscam_session=...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"

  /progress/reset:
    post:
      tags: [Пользователь]
      summary: Сбросить прогресс по роли
      description: |
        Удаляет ответы обучения, закрывает активную сессию экзамена и удаляет сохранённый
        результат. Профиль пользователя не трогает.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: "#/components/schemas/Role"
      responses:
        "200":
          description: Прогресс роли обнулён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /progress/reset/{role}:
    post:
      tags: [Пользователь]
      summary: Сбросить прогресс, роль в пути
      description: Алиас к `POST /progress/reset`, тело не требуется.
      parameters:
        - $ref: "#/components/parameters/RolePath"
      responses:
        "200":
          description: Прогресс роли обнулён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /training/current-step:
    get:
      tags: [Обучение]
      summary: Текущий шаг обучения
      description: |
        Правильный ответ наружу не отдаётся: в `variants` только идентификаторы и тексты.
        Проверка выполняется на сервере в `POST /training/answer`.

        `currentStep` здесь — **1-based номер текущего шага**, в отличие от
        `RoleProgress.training.currentStep`, где это число уже завершённых шагов.
      parameters:
        - $ref: "#/components/parameters/RoleQuery"
      responses:
        "200":
          description: Шаг обучения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingStep"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Обучение по этой роли уже пройдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: TRAINING_ALREADY_PASSED
                  message: Обучение по этой роли уже пройдено

  /training/answer:
    post:
      tags: [Обучение]
      summary: Ответ на шаг обучения
      description: |
        Сохранение ответа и сдвиг указателя выполняются одной транзакцией.

        `stepNumber` необязателен. Если прислан и не совпадает с текущим шагом — ответ
        не записывается, указатель не двигается, возвращается `STEP_MISMATCH` с актуальным
        состоянием в `details`. Так гасится двойной клик и повтор запроса.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: "#/components/schemas/Role"
                answer_id:
                  type: integer
                  format: int64
                  description: Единственное snake_case-поле в контракте
                  example: 42
                answerId:
                  type: integer
                  format: int64
                  description: Синоним `answer_id`, принимается наравне
                stepNumber:
                  type: integer
                  description: Необязателен, защита от повторной отправки
                  example: 1
      responses:
        "200":
          description: Ответ засчитан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Шаг не совпадает с текущим
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: STEP_MISMATCH
                  message: Шаг не совпадает с текущим

  /exam/start:
    get:
      tags: [Экзамен]
      summary: Начать или восстановить экзамен
      description: |
        Если активная сессия уже есть — возвращается она целиком вместе с историей сообщений,
        новая не создаётся и модель не дёргается. Именно так чат восстанавливается после
        перезагрузки страницы. Начать заново — только через `POST /exam/restart`.
      parameters:
        - $ref: "#/components/parameters/RoleQuery"
      responses:
        "200":
          description: Состояние экзамена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Обучение по роли ещё не пройдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: TRAINING_NOT_PASSED
                  message: Сначала нужно пройти обучение по этой роли
        "503":
          $ref: "#/components/responses/LLMUnavailable"
    post:
      tags: [Экзамен]
      summary: То же, роль в теле
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleBody"
      responses:
        "200":
          description: Состояние экзамена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamState"

  /exam/restart:
    post:
      tags: [Экзамен]
      summary: Начать экзамен заново
      description: Закрывает активную сессию и создаёт новую с чистой историей.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleBody"
      responses:
        "200":
          description: Новая сессия
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/LLMUnavailable"

  /exam/message:
    post:
      tags: [Экзамен]
      summary: Сообщение в диалоге
      description: |
        Реплика пользователя проходит через классификатор — отдельный вызов модели, который
        возвращает строгий JSON. Решение о вердикте принимает код сервиса, а не диалоговая
        модель.

        Экзамен завершается по исходу разговора, а не по счётчику ходов: критическая ошибка,
        твёрдый отказ с выходом из разговора, исчерпание всех приёмов мошенника либо
        достижение `maxCycles` как страховка.

        Ограничения: 1000 символов на сообщение, 20 сообщений в минуту на сессию.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, text]
              properties:
                role:
                  $ref: "#/components/schemas/Role"
                text:
                  type: string
                  maxLength: 1000
                  example: Ничего переводить не буду, только через площадку.
      responses:
        "200":
          description: Ответ собеседника
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamReply"
              examples:
                продолжается:
                  value:
                    message: Понимаю осторожность, но давайте быстрее.
                    isFinished: false
                    verdict: null
                    explanation: null
                    cycle: 3
                    maxCycles: 8
                завершён:
                  value:
                    message: Жаль, что не договорились.
                    isFinished: true
                    verdict: passed
                    explanation: Собеседник перепробовал все свои приёмы, и ни один не сработал.
                    cycle: 7
                    maxCycles: 8
        "400":
          description: Пустое или слишком длинное сообщение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Активной сессии нет
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Слишком много сообщений
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          $ref: "#/components/responses/LLMUnavailable"

  /exam/finish:
    post:
      tags: [Экзамен]
      summary: Завершить разговор досрочно
      description: |
        Добровольный выход без критических ошибок засчитывается как успех. Выход до того, как
        собеседник успел применить хоть один приём, тоже даёт `passed`, но половинный балл за
        экзамен: проверки фактически не было.

        Вызов обязателен для записи результата — без него `GET /results` останется пустым.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleBody"
      responses:
        "200":
          description: Вердикт
          content:
            application/json:
              schema:
                type: object
                properties:
                  verdict:
                    $ref: "#/components/schemas/Verdict"
                  explanation:
                    type: string
                  isFinished:
                    type: boolean
                  cycle:
                    type: integer
                  maxCycles:
                    type: integer
        "404":
          description: Активной сессии нет
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /results:
    get:
      tags: [Результаты]
      summary: Итоги по роли
      description: |
        Результат читается из хранилища, а не пересчитывается: он записывается один раз в момент
        завершения экзамена.

        `strengths` и `weaknesses` персональные — считаются по реальному диалогу. Детерминированная
        часть берёт приёмы, которые применил мошенник, и вычитает те, на которых пользователь
        споткнулся; текст пишет модель. При её недоступности разбор собирается из одних фактов.
      parameters:
        - $ref: "#/components/parameters/RoleQuery"
      responses:
        "200":
          description: Итоги
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Results"
        "404":
          description: Экзамен ещё не завершён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: RESULTS_NOT_READY
                  message: Результаты появятся после завершения экзамена
    post:
      tags: [Результаты]
      summary: То же, роль в теле
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleBody"
      responses:
        "200":
          description: Итоги
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Results"
        "404":
          description: Экзамен ещё не завершён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /healthz:
    get:
      tags: [Служебные]
      summary: Проверка живости сервиса и базы
      responses:
        "200":
          description: Сервис в порядке
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "503":
          description: База недоступна
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  parameters:
    RoleQuery:
      name: role
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/Role"
    RolePath:
      name: role
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Role"

  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Куки нет или пользователь не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: USER_NOT_FOUND
              message: Пользователь не найден, создайте профиль
    LLMUnavailable:
      description: Модель недоступна
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: LLM_UNAVAILABLE
              message: Собеседник временно недоступен, попробуйте ещё раз

  schemas:
    Role:
      type: string
      enum: [buyer, seller]

    Gender:
      type: string
      enum: [male, female]

    Verdict:
      type: string
      enum: [passed, failed]
      nullable: true

    Status:
      type: string
      description: |
        По статусу клиент решает, какую кнопку показать, не собирая состояние сам.
      enum:
        - not_started
        - training_in_progress
        - training_passed
        - exam_in_progress
        - exam_passed
        - exam_failed

    RoleBody:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/Role"

    RoleProgress:
      type: object
      required: [training, isExamPassed]
      properties:
        training:
          type: object
          required: [currentStep, totalSteps]
          properties:
            currentStep:
              type: integer
              description: |
                Число **завершённых** шагов: 0 — обучение не начато, totalSteps — пройдено.
                Не путать с `currentStep` в ответе `/training/current-step`.
              example: 2
            totalSteps:
              type: integer
              example: 6
        isExamPassed:
          type: boolean
        status:
          $ref: "#/components/schemas/Status"
        isTrainingPassed:
          type: boolean

    User:
      type: object
      properties:
        name:
          type: string
        age:
          type: string
        gender:
          $ref: "#/components/schemas/Gender"
        buyer:
          $ref: "#/components/schemas/RoleProgress"
        seller:
          $ref: "#/components/schemas/RoleProgress"

    MeResponse:
      type: object
      required: [exists, user]
      properties:
        exists:
          type: boolean
        user:
          allOf:
            - $ref: "#/components/schemas/User"
          nullable: true

    TrainingVariant:
      type: object
      required: [id, text]
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string

    TrainingStep:
      type: object
      required: [currentStep, totalSteps, productName, message, variants]
      properties:
        currentStep:
          type: integer
          description: 1-based номер текущего шага
          example: 1
        totalSteps:
          type: integer
          example: 6
        productName:
          type: string
          example: iPhone 13, 128 ГБ
        message:
          type: string
          description: Реплика наставника с описанием ситуации
        variants:
          type: array
          items:
            $ref: "#/components/schemas/TrainingVariant"

    AnswerResult:
      type: object
      required: [isCorrect, explanation]
      properties:
        isCorrect:
          type: boolean
        explanation:
          type: string
          description: Единый разбор шага, подходит и к верному, и к неверному ответу
        correctId:
          type: integer
          format: int64
        currentStep:
          type: integer
          description: Уже сдвинутый указатель — число завершённых шагов
        totalSteps:
          type: integer
        isTrainingFinished:
          type: boolean

    ExamMessage:
      type: object
      properties:
        id:
          type: integer
          format: int64
        author:
          type: string
          enum: [scammer, user]
        text:
          type: string
        createdAt:
          type: string
          format: date-time

    ExamState:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Последняя реплика собеседника
        sessionId:
          type: string
          format: uuid
        messages:
          type: array
          description: Полная история — по ней восстанавливается чат после перезагрузки
          items:
            $ref: "#/components/schemas/ExamMessage"
        isFinished:
          type: boolean
        verdict:
          $ref: "#/components/schemas/Verdict"
        explanation:
          type: string
          nullable: true
        cycle:
          type: integer
          description: Сделано ходов
        maxCycles:
          type: integer
          description: Страховочный потолок; обычно экзамен завершается раньше по исходу

    ExamReply:
      type: object
      required: [message, isFinished, verdict, explanation]
      properties:
        message:
          type: string
        isFinished:
          type: boolean
        verdict:
          $ref: "#/components/schemas/Verdict"
        explanation:
          type: string
          nullable: true
        cycle:
          type: integer
        maxCycles:
          type: integer

    Results:
      type: object
      required: [role, training, exam, tips]
      properties:
        role:
          $ref: "#/components/schemas/Role"
        training:
          type: object
          properties:
            correctSteps:
              type: integer
            totalSteps:
              type: integer
            answers:
              type: array
              items:
                type: object
                properties:
                  stepNumber:
                    type: integer
                  isCorrect:
                    type: boolean
        exam:
          type: object
          properties:
            verdict:
              $ref: "#/components/schemas/Verdict"
            explanation:
              type: string
            cyclesPassed:
              type: integer
            criticalMistakes:
              type: array
              items:
                type: string
                enum:
                  - leaked_sms_code
                  - leaked_card_data
                  - agreed_to_pay_via_link
                  - agreed_off_platform
            endReason:
              type: string
              enum:
                - critical_mistake
                - refused_and_ended
                - tactics_exhausted
                - user_finished
                - limit_reached
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: 60 × (верных шагов / всего) + 40 × оценка экзамена
        grade:
          type: string
          enum: [Новичок, Осторожный, Уверенный, Эксперт]
        strengths:
          type: array
          description: Что пользователь сделал верно — по реальному диалогу
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        tips:
          type: array
          items:
            type: string

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - USER_NOT_FOUND
                - TRAINING_NOT_PASSED
                - TRAINING_ALREADY_PASSED
                - STEP_MISMATCH
                - INVALID_OPTION
                - SESSION_NOT_FOUND
                - SESSION_ALREADY_FINISHED
                - MESSAGE_TOO_LONG
                - RATE_LIMITED
                - RESULTS_NOT_READY
                - LLM_UNAVAILABLE
                - BAD_REQUEST
                - INTERNAL
            message:
              type: string
            details:
              description: Для STEP_MISMATCH — актуальное состояние прогресса